# meal-planner-system-frontend/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
        "/api/ingredients/{id}": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "getById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngredientResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateIngredientRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngredientResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/ingredients/{id}/dietary-tags": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "getTagsByIngredient",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/IngredientDietaryTagResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "updateTags",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateDietaryTagsRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/ingredients": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "getAll",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "available",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "categoryId",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PagedModelIngredientSummaryResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "create",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateIngredientRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngredientResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/recipes/slug/{slug}": {
            "get": {
                "tags": [
                    "Recipes"
                ],
                "operationId": "getBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/RecipeResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/ingredients/slug/{slug}": {
            "get": {
                "tags": [
                    "Ingredients"
                ],
                "operationId": "getBySlug_1",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/IngredientResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/categories/slug/{slug}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "operationId": "getBySlug_2",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "UpdateIngredientRequest": {
                "type": "object",
                "properties": {
                    "normalizedName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "available": {
                        "type": "boolean"
                    },
                    "price": {
                        "type": "number"
                    },
                    "unit": {
                        "type": "string",
                        "enum": [
                            "KG",
                            "G",
                            "L",
                            "ML",
                            "PCS",
                            "BUNCH"
                        ]
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "aliases": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "categoryId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "normalizedName"
                ]
            },
            "IngredientResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "normalizedName": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "unit": {
                        "type": "string",
                        "enum": [
                            "KG",
                            "G",
                            "L",
                            "ML",
                            "PCS",
                            "BUNCH"
                        ]
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "available": {
                        "type": "boolean"
                    },
                    "categoryId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "categoryName": {
                        "type": "string"
                    },
                    "aliases": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "UpdateDietaryTagsRequest": {
                "type": "object",
                "properties": {
                    "tags": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string",
                            "enum": [
                                "ALLOWED",
                                "SOFT_FORBIDDEN",
                                "HARD_FORBIDDEN"
                            ]
                        }
                    }
                }
            },
...
            "CategoryResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "parentId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "parentName": {
                        "type": "string"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CategoryResponse"
                        }
                    }
                }
            },
...
            "CreateIngredientRequest": {
                "type": "object",
                "properties": {
                    "normalizedName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "available": {
                        "type": "boolean"
                    },
                    "price": {
                        "type": "number"
                    },
                    "unit": {
                        "type": "string",
                        "enum": [
                            "KG",
                            "G",
                            "L",
                            "ML",
                            "PCS",
                            "BUNCH"
                        ]
                    },
                    "stock": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "aliases": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "categoryId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "normalizedName"
                ]
            },
...
            "PageMetadata": {
                "type": "object",
                "properties": {
                    "size": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "number": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "RecipeIngredientResponse": {
                "type": "object",
                "properties": {
                    "ingredientId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "rawName": {
                        "type": "string"
                    },
                    "rawAmount": {
                        "type": "string"
                    },
                    "availableInShop": {
                        "type": "boolean"
                    },
                    "price": {
                        "type": "number"
                    },
                    "unit": {
                        "type": "string",
                        "enum": [
                            "KG",
                            "G",
                            "L",
                            "ML",
                            "PCS",
                            "BUNCH"
                        ]
                    },
                    "imageUrl": {
                        "type": "string"
                    }
                }
            },
            "RecipeResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "mealType": {
                        "type": "string"
                    },
                    "mealTypeDetailed": {
                        "type": "string"
                    },
                    "cookTime": {
                        "type": "string"
                    },
                    "cookComplexity": {
                        "type": "string"
                    },
                    "cookBudget": {
                        "type": "string"
                    },
                    "servings": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "servingSize": {
                        "type": "string"
                    },
                    "steps": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "ingredients": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RecipeIngredientResponse"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 0
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32",
                        "minimum": 1
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "IngredientSummaryResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "normalizedName": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "unit": {
                        "type": "string",
                        "enum": [
                            "KG",
                            "G",
                            "L",
                            "ML",
                            "PCS",
                            "BUNCH"
                        ]
                    },
                    "available": {
                        "type": "boolean"
                    },
                    "categoryId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "categoryName": {
                        "type": "string"
                    }
                }
            },
            "PagedModelIngredientSummaryResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IngredientSummaryResponse"
                        }
                    },
                    "page": {
                        "$ref": "#/components/schemas/PageMetadata"
                    }
                }
            },
            "IngredientDietaryTagResponse": {
                "type": "object",
                "properties": {
                    "conditionId": {
                        "type": "string"
                    },
                    "conditionName": {
                        "type": "string"
                    },
                    "conditionType": {
                        "type": "string",
                        "enum": [
                            "CONTRAINDICATION",
                            "DIET"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "ALLOWED",
                            "SOFT_FORBIDDEN",
                            "HARD_FORBIDDEN"
                        ]
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}